Combined Sup3rGan and WindGan to remove lots of duplicate code. Sup3rGan/Sup3rCondMom/Sup3rGanDC now works with/without mid network exo injection. Transpose method added to MultiStepGan to match input data shape with model.input_dims. This should accomplish the same as SpatialThenTemporal/TemporalThenSpatial models in addition to a st model + spatial + st model, for example.
