+++ sup3r/models/abstract.py
@@ -121,8 +121,9 @@ class AbstractInterface(ABC):
                                           str(output_spatial))
         return output_res
 
-    def _combine_input(self, low_res, exogenous_data=None):
-        """Combine exogenous_data at input resolution with low_res data
+    def _combine_fwp_input(self, low_res, exogenous_data=None):
+        """Combine exogenous_data at input resolution with low_res data prior
+        to forward pass through generator
 
         Parameters
         ----------
@@ -166,8 +167,9 @@ class AbstractInterface(ABC):
                                              axis=-1)
         return low_res
 
-    def _combine_output(self, hi_res, exogenous_data=None):
-        """Combine exogenous_data at input resolution with low_res data
+    def _combine_fwp_output(self, hi_res, exogenous_data=None):
+        """Combine exogenous_data at output resolution with generated hi_res
+        data following forward pass output.
 
         Parameters
         ----------
@@ -209,6 +211,50 @@ class AbstractInterface(ABC):
                                             axis=-1)
         return hi_res
 
+    def _combine_loss_input(self, high_res_true, high_res_gen):
+        """Combine exogenous feature data from high_res_true with high_res_gen
+        for loss calculation
+
+        Parameters
+        ----------
+        high_res_true : tf.Tensor
+            Ground truth high resolution spatiotemporal data.
+        high_res_gen : tf.Tensor
+            Superresolved high resolution spatiotemporal data generated by the
+            generative model.
+
+        Returns
+        -------
+        high_res_gen : tf.Tensor
+            Same as input with exogenous data combined with high_res input
+        """
+        for feature in self.exogenous_features:
+            f_idx = self.training_features.index(feature)
+            exo_data = high_res_true[..., f_idx: f_idx + 1]
+            high_res_gen = tf.concat((high_res_gen, exo_data), axis=-1)
+        return high_res_gen
+
+    def _get_exo_val_loss_input(self, high_res):
+        """Get exogenous feature data from high_res
+
+        Parameters
+        ----------
+        high_res : tf.Tensor
+            Ground truth high resolution spatiotemporal data.
+
+        Returns
+        -------
+        exo_data : dict
+            Dictionary of exogenous feature data used as input to tf_generate.
+            e.g. {'topography': np.ndarray(...)}
+        """
+        exo_data = {}
+        for feature in self.exogenous_features:
+            f_idx = self.training_features.index(feature)
+            exo_fdata = high_res[..., f_idx: f_idx + 1]
+            exo_data[feature] = exo_fdata
+        return exo_data
+
     @property
     def exogenous_features(self):
         """Get list of exogenous filter names the model uses. If the model has
@@ -219,12 +265,12 @@ class AbstractInterface(ABC):
         [topo, sza]. Topo will then be used in the first concat layer and sza
         will be used in the second"""
         # pylint: disable=E1101
-        layer_count = 0
+        features = []
         if hasattr(self, '_gen'):
-            layer_count = sum(
-                isinstance(layer, (Sup3rAdder, Sup3rConcat))
-                for layer in self._gen.layers)
-        return self.training_features[-layer_count:]
+            for layer in self._gen.layers:
+                if isinstance(layer, (Sup3rAdder, Sup3rConcat)):
+                    features.append(layer.name)
+        return features
 
     @property
     @abstractmethod
@@ -277,6 +323,35 @@ class AbstractInterface(ABC):
         """
         return VERSION_RECORD
 
+    def _check_exo_features(self, **kwargs):
+        """Make sure exogenous features have the correct ordering and are
+        included in training_features
+
+        Parameters
+        ----------
+        kwargs : dict
+            Keyword arguments including 'training_features', 'output_features',
+            'smoothed_features', 's_enhance', 't_enhance', 'smoothing'
+
+        Returns
+        -------
+        kwargs : dict
+            Same as input but with exogenous_features removed from output
+            features
+        """
+        output_features = kwargs['output_features']
+        msg = (f'Last {len(self.exogenous_features)} output features from the '
+               f'data handler must be {self.exogenous_features} '
+               'to train the Exo model, but received output features: {}'.
+               format(output_features))
+        check = (output_features[-len(self.exogenous_features)]
+                 == self.exogenous_features)
+        assert check, msg
+        for f in self.exogenous_features:
+            output_features.remove(f)
+        kwargs['output_features'] = output_features
+        return kwargs
+
     def set_model_params(self, **kwargs):
         """Set parameters used for training the model
 
@@ -286,6 +361,7 @@ class AbstractInterface(ABC):
             Keyword arguments including 'training_features', 'output_features',
             'smoothed_features', 's_enhance', 't_enhance', 'smoothing'
         """
+        kwargs = self._check_exo_features(**kwargs)
 
         keys = ('input_resolution', 'training_features', 'output_features',
                 'smoothed_features', 's_enhance', 't_enhance', 'smoothing')
@@ -1090,7 +1166,7 @@ class AbstractSingleModel(ABC):
             (n_obs, spatial_1, spatial_2, n_features)
             (n_obs, spatial_1, spatial_2, n_temporal, n_features)
         """
-        low_res = self._combine_input(low_res, exogenous_data)
+        low_res = self._combine_fwp_input(low_res, exogenous_data)
 
         if norm_in and self._means is not None:
             low_res = self.norm_input(low_res)
@@ -1101,10 +1177,12 @@ class AbstractSingleModel(ABC):
                 if isinstance(layer, (Sup3rAdder, Sup3rConcat)):
                     exo_name = layer.name
                     steps = exogenous_data[exo_name]['steps']
-                    idx = [step['combine_type'] for step in steps]
+                    combine_types = [step['combine_type'] for step in steps]
+                    idx = combine_types.index('layer')
                     hi_res_exo = steps[idx]['data']
                     hi_res_exo = self._reshape_norm_exo(hi_res,
                                                         hi_res_exo,
+                                                        exo_name,
                                                         norm_in=norm_in)
                     hi_res = layer(hi_res, hi_res_exo)
                 else:
@@ -1120,7 +1198,7 @@ class AbstractSingleModel(ABC):
         if un_norm_out and self._means is not None:
             hi_res = self.un_norm_output(hi_res)
 
-        hi_res = self._combine_output(hi_res, exogenous_data)
+        hi_res = self._combine_fwp_output(hi_res, exogenous_data)
 
         return hi_res
 
@@ -1222,44 +1300,3 @@ class AbstractSingleModel(ABC):
                 grad = tape.gradient(loss, training_weights)
 
         return grad, loss_details
-
-
-# pylint: disable=E1101,W0201,E0203
-class AbstractExoInterface(AbstractInterface):
-    """
-    Abstract class to define the required training interface
-    for Sup3r model subclasses with exogenous features
-    """
-
-    # pylint: disable=E0211
-    def set_model_params(self, **kwargs):
-        """Set parameters used for training the model
-
-        Parameters
-        ----------
-        kwargs : dict
-            Keyword arguments including 'input_resolution',
-            'training_features', 'output_features', 'smoothed_features',
-            's_enhance', 't_enhance', 'smoothing'. For the Wind classes, the
-            last entry in "output_features" must be "topography"
-
-        Returns
-        -------
-        kwargs : dict
-            Same as input but with exogenous features removed from
-            "output_features", this is because the exo features are
-            concatenated mid-network in the ExoGan generators and are not
-            output features but are required in the hi-res training set.
-        """
-        output_features = kwargs['output_features']
-        msg = (f'Last {len(self.exogenous_features)} output features from the '
-               f'data handler must be {self.exogenous_features} '
-               'to train the Exo model, but received output features: {}'.
-               format(output_features))
-        check = (output_features[-len(self.exogenous_features)]
-                 == self.exogenous_features)
-        assert check, msg
-        for f in self.exogenous_features:
-            output_features.remove(f)
-        kwargs['output_features'] = output_features
-        return kwargs
