+++ sup3r/models/data_centric.py
@@ -6,7 +6,6 @@ import logging
 import numpy as np
 
 from sup3r.models.base import Sup3rGan
-from sup3r.models.multi_exo import MultiExoGan
 from sup3r.utilities.utilities import round_array
 
 logger = logging.getLogger(__name__)
@@ -39,7 +38,8 @@ class Sup3rGanDC(Sup3rGan):
         """
         losses = []
         for obs in batch_handler.val_data:
-            gen = self._tf_generate(obs.low_res)
+            exo_data = self._get_exo_val_loss_input(obs.high_res)
+            gen = self._tf_generate(obs.low_res, exo_data)
             loss, _ = self.calc_loss(obs.high_res, gen,
                                      weight_gen_advers=weight_gen_advers,
                                      train_gen=True, train_disc=True)
@@ -66,7 +66,8 @@ class Sup3rGanDC(Sup3rGan):
         """
         losses = []
         for obs in batch_handler.val_data:
-            gen = self._tf_generate(obs.low_res)
+            exo_data = self._get_exo_val_loss_input(obs.high_res)
+            gen = self._tf_generate(obs.low_res, exo_data)
             loss = self.calc_loss_gen_content(obs.high_res, gen)
             losses.append(float(loss))
         return losses
@@ -93,7 +94,6 @@ class Sup3rGanDC(Sup3rGan):
             Updated loss_details with mean validation loss calculated using
             the validation samples across the time bins
         """
-
         total_losses = self.calc_val_loss_gen(batch_handler, weight_gen_advers)
         content_losses = self.calc_val_loss_gen_content(batch_handler)
 
@@ -136,71 +136,6 @@ class Sup3rGanDC(Sup3rGan):
                      f'{round_array(t_c_losses)}')
         logger.info('Updated temporal bin weights: '
                     f'{round_array(new_temporal_weights)}')
-
-
-class MultiExoGanDC(MultiExoGan, Sup3rGanDC):
-    """Data-centric model using loss across time bins to select training
-    observations with handling of low and high res topography
-    inputs."""
-
-    def calc_val_loss_gen(self, batch_handler, weight_gen_advers):
-        """Calculate the validation total loss across the validation
-        samples. e.g. If the sample domain has 100 steps and the
-        validation set has 10 bins then this will get a list of losses across
-        step 0 to 10, 10 to 20, etc.  Use this to determine performance
-        within bins and to update how observations are selected from these
-        bins. Use the _tf_generate function from MultiExoGan to include the
-        high resolution topography.
-
-        Parameters
-        ----------
-        batch_handler : sup3r.data_handling.preprocessing.BatchHandlerDC
-            BatchHandler object to iterate through
-        weight_gen_advers : float
-            Weight factor for the adversarial loss component of the generator
-            vs. the discriminator.
-
-        Returns
-        -------
-        list
-            List of total losses for all sample bins
-        """
-        losses = []
-        for obs in batch_handler.val_data:
-            gen = self._tf_generate(obs.low_res,
-                                    obs.high_res[..., -1:])
-            loss, _ = self.calc_loss(obs.high_res, gen,
-                                     weight_gen_advers=weight_gen_advers,
-                                     train_gen=True, train_disc=True)
-            losses.append(float(loss))
-        return losses
-
-    def calc_val_loss_gen_content(self, batch_handler):
-        """Calculate the validation content loss across the validation
-        samples. e.g. If the sample domain has 100 steps and the
-        validation set has 10 bins then this will get a list of losses across
-        step 0 to 10, 10 to 20, etc.  Use this to determine performance
-        within bins and to update how observations are selected from these
-        bins. Use the _tf_generate function from MultiExoGan to include high
-        resolution topography.
-
-        Parameters
-        ----------
-        batch_handler : sup3r.data_handling.preprocessing.BatchHandlerDC
-            BatchHandler object to iterate through
-
-        Returns
-        -------
-        list
-            List of content losses for all sample bins
-        """
-        losses = []
-        for obs in batch_handler.val_data:
-            gen = self._tf_generate(obs.low_res,
-                                    obs.high_res[..., -1:])
-            loss = self.calc_loss_gen_content(obs.high_res, gen)
-            losses.append(float(loss))
-        return losses
 
 
 class Sup3rGanSpatialDC(Sup3rGanDC):
